pyproject.toml
[tool.poetry]
package-mode = false
name = "cyber_wizard"
version = "0.1.0"
description = ""
authors = ["Irishery <markirill21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.11.0.86"
requests = "^2.32.3"
speechrecognition = "^3.14.1"
pyaudio = "^0.2.14"
openai = "^1.65.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
